Initial Commit of Basic Code
Currently checking in the proof of concept code. The Arduino will respond to the GetPage request from the client C# application.
The C# App will add a button and a text box to the canvas. for some reason currently the arduino is sending a 63 instead of 200 for the height.. not sure why that is. But that is where I left off. Sending 100,100 works fine. But 200 gets changed to 63 for some reason